complete 4.21 bloglist expansion step9

+++ part4/blog-list/controllers/blogs.js
@@ -1,8 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
 const jwt = require('jsonwebtoken')
+
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -24,9 +32,10 @@ blogsRouter.get('/:id', (request, response, next) => {
 blogsRouter.post('/', async (request, response) => {
   //const user = await User.findOne();
   const body = request.body;
-  const token = request.token;
-
-  const decodedToken = jwt.verify(token, process.env.SECRET)
+  //const token = request.token;
+  
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
@@ -52,12 +61,24 @@ blogsRouter.post('/', async (request, response) => {
 
   user.blogs = user.blogs.concat(savedBlog._id) // Update the user's 'blogs' array
   await user.save()
+  response.status(201).json(savedBlog.toJSON());
 })
 
+
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id);
+
+  if ( blog.user.toString() === user.id.toString() )
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
+
 
 blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
